multimodal-understanding/workshop/2-multimodal-understanding/requirements.txt
# Package management and build tools
pip>=23.0.0
setuptools>=68.0.0
packaging>=23.0.0

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# Image and Video Processing
opencv-python-headless>=4.12.0
scikit-image>=0.25.2
Pillow>=11.3.0

# Numerical and Scientific Computing
numpy>=1.26.4  # Version specified to match Sagemaker requirement

# Jupyter/IPython Environment
IPython>=8.0.0
